Move Boundary Lookups Into modeling/calcs.py
- modeling/calcs.py was entirely unused, and contained
  the same code as geoprocessing_api/calcs.py

- In preparation for moving AOI/WKAOI parsing out into
  a serializer, move boundary lookups and calcs from
  modeling/views to modeling/calcs, where both the
  future serializer and the views will have access
