+++ src/mmw/apps/modeling/views.py
@@ -17,9 +17,7 @@ from rest_framework.permissions import (AllowAny,
 
 from django.shortcuts import get_object_or_404
 from django.utils.timezone import now
-from django.conf import settings
 from django.db import connection
-from django.contrib.gis.geos import WKBReader
 from django.http import (HttpResponse,
                          Http404,
                          )
@@ -37,6 +35,8 @@ from apps.modeling.serializers import (ProjectSerializer,
                                        ProjectListingSerializer,
                                        ProjectUpdateSerializer,
                                        ScenarioSerializer)
+from apps.modeling.calcs import (get_layer_shape,
+                                 boundary_search_context)
 
 
 @decorators.api_view(['GET', 'POST'])
@@ -332,13 +332,6 @@ def _construct_tr55_job_chain(model_input, job_id):
     return job_chain
 
 
-def _get_boundary_layer_by_code(code):
-    for layer in settings.LAYER_GROUPS['boundary']:
-        if layer.get('code') == code:
-            return layer
-    return False
-
-
 @decorators.api_view(['GET'])
 @decorators.permission_classes((AllowAny, ))
 def boundary_layer_detail(request, table_code, obj_id):
@@ -350,87 +343,6 @@ def boundary_layer_detail(request, table_code, obj_id):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
 
-def _get_boundary_search_query(search_term):
-    """
-    Return raw SQL query to perform full text search against
-    all searchable boundary layers.
-    """
-    select_fmt = """
-        (SELECT id, '{code}' AS code, name, {rank} AS rank,
-            ST_Centroid(geom) as center
-        FROM {table}
-        WHERE UPPER(name) LIKE UPPER(%(term)s)
-        LIMIT 3)
-    """.strip()
-
-    selects = []
-    for layer in settings.LAYER_GROUPS['boundary']:
-        if not layer.get('searchable'):
-            continue
-
-        code = layer['code']
-        table_name = layer['table_name']
-        rank = layer.get('search_rank', 0)
-
-        selects.append(select_fmt.format(
-            code=code, table=table_name, rank=rank))
-
-    if len(selects) == 0:
-        raise Exception('No boundary layers are searchable')
-
-    subquery = ' UNION ALL '.join(selects)
-
-    return """
-        SELECT id, code, name, rank, center
-        FROM ({}) AS subquery
-        ORDER BY rank DESC, name
-    """.format(subquery)
-
-
-def _do_boundary_search(search_term):
-    """
-    Execute full text search against all searchable boundary layers.
-    """
-    result = []
-    query = _get_boundary_search_query(search_term)
-
-    with connection.cursor() as cursor:
-        wildcard_term = '%{}%'.format(search_term)
-        cursor.execute(query, {'term': wildcard_term})
-
-        wkb_r = WKBReader()
-
-        for row in cursor.fetchall():
-            id = row[0]
-            code = row[1]
-            name = row[2]
-            rank = row[3]
-            point = wkb_r.read(row[4])
-
-            layer = _get_boundary_layer_by_code(code)
-
-            result.append({
-                'id': id,
-                'code': code,
-                'text': name,
-                'label': layer['short_display'],
-                'rank': rank,
-                'y': point.y,
-                'x': point.x,
-            })
-
-    return result
-
-
-def _boundary_search_context(search_term):
-    suggestions = [] if len(search_term) < 3 else \
-        _do_boundary_search(search_term)
-    # Data format should match the ArcGIS API suggest endpoint response
-    return {
-        'suggestions': suggestions,
-    }
-
-
 @decorators.api_view(['GET'])
 @decorators.permission_classes((AllowAny, ))
 def boundary_layer_search(request):
@@ -438,7 +350,7 @@ def boundary_layer_search(request):
     if not search_term:
         return Response('Missing query string param: text',
                         status=status.HTTP_400_BAD_REQUEST)
-    result = _boundary_search_context(search_term)
+    result = boundary_search_context(search_term)
     return Response(result)
 
 
@@ -552,37 +464,6 @@ def get_job(request, job_uuid, format=None):
             'finished': job.delivered_at,
         }
     )
-
-
-def get_layer_shape(table_code, id):
-    """
-    Fetch shape of well known area of interest.
-
-    :param table_code: Code of table
-    :param id: ID of the shape
-    :return: GeoJSON of shape if found, None otherwise
-    """
-    layer = _get_boundary_layer_by_code(table_code)
-    if not layer:
-        return None
-
-    table = layer['table_name']
-    field = layer.get('json_field', 'geom')
-
-    sql = '''
-          SELECT ST_AsGeoJSON({field})
-          FROM {table}
-          WHERE id = %s
-          '''.format(field=field, table=table)
-
-    with connection.cursor() as cursor:
-        cursor.execute(sql, [int(id)])
-        row = cursor.fetchone()
-
-        if row:
-            return row[0]
-        else:
-            return None
 
 
 def parse_area_of_interest(area_of_interest):
